Special Case permitting single row to referenced in Range()
A special case where only the header row needs to be referenced appears difficult to implement. Referencing the table header row is useful in e.g. Index(Match()) type formulae. 

Suggest a special case where bottom_row=False (rather than None or a index item key being provided) in which the bottom_row_offset is set to the top_row_offset so as to reference the single header row.